Filter closed restaurants in talabat query
Filter out and remove any restaurant that is closed in talabat. this filteration step occurs while filtering for deals (before requesting menu items).

since most of the restaurants returned from the query are actually closed, this has dramatically affected number of returned restaurants and request speed! for my current location, it returned 3 instead of 21 restaurants, and the speed increase has been proportional to the loss of restaurants.

Now the results are a bit sparse and empty looking, but moving forward i think it's best to display some statistics about "pruned" restaurants. the list will be longer once more APIs are scraped as well