Implement change of response data
Changed the frontend's processing of the data returned from the query function. originally it returned an array of restaurants. now it returns a hash of information, including the restaurant array.

Refactored the API, removed the API component specifically for talabat and migrated it's functions to the main API aggregator. this is because the aggregation will happen server-side so there is no need for the client to call multiple points for different providers.

Now shows simple information above the list of restaurant data, includes Number of pruned restaurants